+++ src/rendering/renderer.h
@@ -6,7 +6,7 @@
 void rend_InitRendering(void);
 void rend_EndRendering(void);
 void rend_PrintMainMenu(void);
-void rend_RenderGame(struct GameState *state, double playerHeight);
+void rend_RenderGame(struct GameState *state, float playerHeight);
 
 #endif /* RENDERER_H */
